code cleanup
git-svn-id: svn+ssh://pal.gootz.net/var/svn/repos/gome@150 ab0d379b-fd29-0410-b982-9e38e41d62e4

+++ sources/com/indigonauts/gome/igs/ServerUser.java
@@ -38,9 +38,9 @@ public class ServerUser extends QuickSortable {
     char second = user.rank.charAt(1);
     char first = user.rank.charAt(0);
     if (!Character.isDigit(second)) {
-      user.numRank = (int) (first - '0');
+      user.numRank = first - '0';
     } else {
-      user.numRank = (int) (10 * (first - '0')) + (second - '0');
+      user.numRank = 10 * (first - '0') + (second - '0');
     }
     return user;
   }
